Merge branch 'dev' into 'master''

Closes CAP-510, CAP-410, and CAP-501

See merge request computational-audition-lab/psynet!74

+++ psynet/media.py
@@ -6,7 +6,9 @@ import botocore.exceptions
 import botocore.errorfactory
 import tempfile
 import shutil
+import numpy as np
 from uuid import uuid4
+from scipy.io import wavfile
 
 from pathlib import Path
 
@@ -97,9 +99,6 @@ def upload_to_s3(local_path: str, bucket_name: str, key: str, public_read: bool,
 
     logger.info("Uploading %s to bucket %s with key %s...", local_path, bucket_name, key)
 
-    # client = new_s3_client()
-    # client.upload_file(local_path, bucket_name, key)
-
     args = {}
     if public_read:
         args["ACL"] = "public-read"
@@ -107,7 +106,18 @@ def upload_to_s3(local_path: str, bucket_name: str, key: str, public_read: bool,
     bucket = get_s3_bucket(bucket_name)
 
     def upload():
-        bucket.upload_file(local_path, key, ExtraArgs=args)
+        if os.path.isfile(local_path):
+            bucket.upload_file(local_path, key, ExtraArgs=args)
+        else:
+            for _dir_path, _dir_names, _file_names in os.walk(local_path):
+                _rel_dir_path = os.path.relpath(_dir_path, local_path)
+                for _file_name in _file_names:
+                    _local_path = os.path.join(_dir_path, _file_name)
+                    if _rel_dir_path == ".":
+                        _file_key = os.path.join(key, _file_name)
+                    else:
+                        _file_key = os.path.join(key, _rel_dir_path, _file_name)
+                    bucket.upload_file(_local_path, _file_key, ExtraArgs=args)
 
     try:
         upload()
@@ -308,7 +318,9 @@ def make_bucket_public(bucket_name):
 def recode_wav(file_path):
     import parselmouth
     with tempfile.NamedTemporaryFile() as temp_file:
-        shutil.copyfile(file_path, temp_file.name)
+        fs, data = wavfile.read(file_path)
+        force_bit_depth = data.astype(np.float32)
+        wavfile.write(temp_file.name, fs, force_bit_depth)
         s = parselmouth.Sound(temp_file.name)
         s.save(file_path, "WAV")
